Move most parsing functions to a separate file
